fix(QuantityInput component): move to components folder
