+++ src/app/sn/carts/Cart/style.module.scss
@@ -12,13 +12,3 @@
   flex-direction: column;
   gap: 10px;
 }
-
-.quantity {
-  width: 110px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 7px 10px;
-  border: 3px solid $pointColor;
-  border-radius: 20px;
-}
